+++ packages/build-info/src/node/get-build-info.ts
@@ -1,6 +1,7 @@
 import { listFrameworks } from '@netlify/framework-info'
 
 import { Logger } from '../file-system.js'
+import { report } from '../metrics.js'
 import { PkgManagerFields } from '../package-managers/detect-package-manager.js'
 import { Project } from '../project.js'
 import { WorkspaceInfo } from '../workspaces/detect-workspace.js'
@@ -28,22 +29,33 @@ export class NoopLogger implements Logger {
   error() {
     /** noop */
   }
+  info() {
+    /** noop */
+  }
+  warn() {
+    /** noop */
+  }
 }
 
 /** Get the build info object that is used inside buildbot */
-export async function getBuildInfo(projectDir?: string, rootDir?: string): Promise<Info> {
+export async function getBuildInfo(
+  config: {
+    projectDir?: string
+    rootDir?: string
+    featureFlags?: Record<string, boolean>
+  } = { featureFlags: {} },
+): Promise<Info> {
   const fs = new NodeFS()
   // prevent logging in output as we use the stdout to capture the json
   fs.logger = new NoopLogger()
-  const project = new Project(fs, projectDir, rootDir)
+  const project = new Project(fs, config.projectDir, config.rootDir)
   project.setEnvironment(process.env)
   let frameworks: any[] = []
   try {
     // if the framework  detection is crashing we should not crash the build info and package-manager detection
     frameworks = await listFrameworks({ projectDir: project.baseDirectory })
-  } catch {
-    // TODO: build reporting to buildbot see: https://github.com/netlify/pillar-workflow/issues/1001
-    // noop
+  } catch (error) {
+    report(error)
   }
 
   const info: Info = {